snowpack_archive/requirements.txt
minio==7.1.13
#pyminio==0.2.1
scandir==1.10.0
python-dotenv==1.0.0
click==8.1.3
